get_channel_id as method of channels class
* the tests make mocking over requests library.
