+++ slacker/__init__.py
@@ -138,6 +138,12 @@ class Channels(BaseAPI):
         return self.get('channels.list',
                         params={'exclude_archived': exclude_archived})
 
+    def get_channel_id(self, channel_name):
+        for channel in self.list().body['channels']:
+            name, channel_id = channel['name'], channel['id']
+            if name == channel_name:
+                return channel_id
+
     def history(self, channel, latest=None, oldest=None, count=None):
         return self.get('channels.history',
                         params={
